+++ src/main/java/com/jonasdurau/ceramicmanagement/services/ResourceService.java
@@ -10,6 +10,7 @@ import com.jonasdurau.ceramicmanagement.dtos.YearReportDTO;
 import com.jonasdurau.ceramicmanagement.entities.Resource;
 import com.jonasdurau.ceramicmanagement.entities.ResourceTransaction;
 import com.jonasdurau.ceramicmanagement.entities.enums.TransactionType;
+import com.jonasdurau.ceramicmanagement.repositories.BatchResourceUsageRepository;
 import com.jonasdurau.ceramicmanagement.repositories.ResourceRepository;
 import com.jonasdurau.ceramicmanagement.repositories.ResourceTransactionRepository;
 
@@ -36,6 +37,9 @@ public class ResourceService {
 
     @Autowired
     private ResourceTransactionRepository transactionRepository;
+
+    @Autowired
+    private BatchResourceUsageRepository batchResourceUsageRepository;
 
     @Transactional(readOnly = true)
     public List<ResourceListDTO> findAll() {
@@ -93,9 +97,12 @@ public class ResourceService {
         Resource entity = resourceRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado. Id: " + id));
         boolean hasTransactions = transactionRepository.existsByResourceId(id);
+        boolean hasBatchUsages = batchResourceUsageRepository.existsByResourceId(id);
         if (hasTransactions) {
-            throw new ResourceDeletionException(
-                    "Não é possível deletar o recurso com id " + id + " pois ele tem transações associadas.");
+            throw new ResourceDeletionException("Não é possível deletar o recurso com id " + id + " pois ele tem transações associadas.");
+        }
+        if (hasBatchUsages) {
+            throw new ResourceDeletionException("Não é possível deletar o recurso com id " + id + " pois ele tem bateladas associadas.");
         }
         resourceRepository.delete(entity);
     }
